+++ src/main/groovy/ru/vyarus/gradle/plugin/lib/JavaLibPlugin.groovy
@@ -81,7 +81,7 @@ class JavaLibPlugin implements Plugin<Project> {
             MavenPublication bom = configureBomPublication(project)
             configurePlatform(project, extension, bom)
             configureGradleMetadata(project, extension)
-            configureSigning(project, bom)
+            configureSigning(project, extension, bom)
             addInstallTask(project) {
                 project.logger.warn "INSTALLED $project.group:$bom.artifactId:$project.version"
             }
@@ -99,7 +99,7 @@ class JavaLibPlugin implements Plugin<Project> {
             addJavadocJarTask(project, publication, extension)
             addGroovydocJarTask(project, publication, extension)
             configureGradleMetadata(project, extension)
-            configureSigning(project, publication)
+            configureSigning(project, extension, publication)
             addInstallTask(project) {
                 project.logger.warn "INSTALLED $project.group:$project.name:$project.version"
             }
@@ -305,14 +305,14 @@ class JavaLibPlugin implements Plugin<Project> {
         }
     }
 
-    private void configureSigning(Project project, MavenPublication publication) {
+    private void configureSigning(Project project, JavaLibExtension extension, MavenPublication publication) {
         project.plugins.withType(SigningPlugin) {
             // https://docs.gradle.org/current/userguide/signing_plugin.html#sec:signatory_credentials
             project.signing.sign publication
 
             // snapshots not signed to simplify project usage (no need for cert during local dev)
             project.tasks.withType(Sign).configureEach {
-                it.onlyIf { !project.version.toString().endsWith('SNAPSHOT') }
+                it.onlyIf { extension.signSnapshots || !project.version.toString().endsWith('SNAPSHOT') }
             }
         }
     }